added support for export and import of process instance, fixed handling of subprocesses in cassandra and dynamo persistence addons

+++ api/automatiko-engine-api/src/main/java/io/automatiko/engine/api/workflow/ProcessInstances.java
@@ -24,6 +24,6 @@ public interface ProcessInstances<T> {
 
     Collection<? extends ProcessInstance<T>> findByIdOrTag(ProcessInstanceReadMode mode, String... values);
 
-    Integer size();
+    Long size();
 
 }
